+++ Legacy/Dashboard/DashboardService.cs
@@ -11,7 +11,7 @@ namespace Legacy.Dashboard
         private readonly ILegacyPostingQueryRepository _postingQueryRepository = postingQueryRepository;
         private readonly ILegacyAccountStatusQueryRepository _monthlyAccountStatusRepository = monthlyAccountStatusRepository;
 
-        public Dictionary<int, Dictionary<string, DashboardInformation>> GetDashboards(int userId, List<int> monthsInDashboard, bool allMonthAndYears, List<double> returnRates, int currentAge, int pensionAge)
+        public Dictionary<int, Dictionary<string, DashboardInformation>> GetDashboards(int userId, List<int> monthsInDashboard, bool allMonthAndYears, List<decimal> returnRates, int currentAge, int pensionAge)
         {
             var monthlyIncomeAndExpenses = _postingQueryRepository.GetMonthlyIncomeAndExpenses(userId);
             var monthlyAccountCategorySums = _monthlyAccountStatusRepository.GetAccumulatedCategorySums(userId);
@@ -28,13 +28,13 @@ namespace Legacy.Dashboard
             return dashboards;
         }
 
-        private static DashboardInformation CalculateDashboard(MonthAndYear currentMonthAndYear, List<double> returnRates, int currentAge, int pensionAge, int numMonths, Dictionary<MonthAndYear, Ledger> monthlyLedgers)
+        private static DashboardInformation CalculateDashboard(MonthAndYear currentMonthAndYear, List<decimal> returnRates, int currentAge, int pensionAge, int numMonths, Dictionary<MonthAndYear, Ledger> monthlyLedgers)
         {
             var ledger = CalculateAverageLedger(monthlyLedgers, numMonths, currentMonthAndYear);
             return CalculateDashboard(ledger, currentAge, pensionAge, returnRates);
         }
 
-        private static DashboardInformation CalculateDashboard(Ledger ledger, int currentAge, int pensionAge, List<double> returnRates)
+        private static DashboardInformation CalculateDashboard(Ledger ledger, int currentAge, int pensionAge, List<decimal> returnRates)
         {
             var yearToPension = pensionAge - currentAge;
