+++ Legacy/Dashboard/IDashboardService.cs
@@ -4,6 +4,6 @@ namespace Legacy.Dashboard
 {
     public interface IDashboardService
     {
-        Dictionary<int,Dictionary<string, DashboardInformation>> GetDashboards(int userId, List<int> monthsInDashboard, bool allMonthAndYears, List<double> returnRates, int currentAge, int pensionAge);
+        Dictionary<int,Dictionary<string, DashboardInformation>> GetDashboards(int userId, List<int> monthsInDashboard, bool allMonthAndYears, List<decimal> returnRates, int currentAge, int pensionAge);
     }
 }
