+++ Legacy/Repositories/ILegacyLossRepository.cs
@@ -4,6 +4,6 @@ namespace Legacy.Repositories
 {
     public interface ILegacyLossRepository
     {
-        void UpdateLoss(int userId, MonthAndYear monthAndYear, double loss);
+        void UpdateLoss(int userId, MonthAndYear monthAndYear, decimal loss);
     }
 }
